# backend/core/docs/swagger.yaml
basePath: /
definitions:
  bookings.TicketType:
    enum:
    - STANDARD
    - HALF
    - FREE
    type: string
    x-enum-varnames:
    - TicketTypeStandard
    - TicketTypeHalf
    - TicketTypeFree
...
  handler.CinemaSessionsResponseDTO:
    properties:
      cinema_city:
        type: string
      cinema_id:
        type: integer
      cinema_name:
        type: string
      sessions:
        items:
          $ref: '#/definitions/handler.SessionResponseDTO'
        type: array
    type: object
...
  handler.MovieResponseDTO:
    properties:
      id:
        type: integer
      is_premiere:
        type: boolean
      is_rescreening:
        type: boolean
      poster_url:
        type: string
      title:
        type: string
      tmdb_id:
        type: integer
    type: object
...
  handler.PayRequestDTO:
    properties:
      payment_method:
        type: string
    required:
    - payment_method
    type: object
  handler.PayResponseDTO:
    properties:
      client_secret:
        type: string
      message:
        type: string
    type: object
...
  handler.ReserveRequestDTO:
    properties:
      session_id:
        type: integer
      tickets_request:
        items:
          $ref: '#/definitions/handler.TicketRequest'
        minItems: 1
        type: array
    required:
    - session_id
    - tickets_request
    type: object
  handler.ReserveResponseDTO:
    properties:
      message:
        type: string
      transaction_id:
        type: string
      valor_total_centavos:
        type: integer
    type: object
...
  handler.SeatResponseDTO:
    properties:
      id:
        type: integer
      isOccupied:
        type: boolean
      number:
        type: integer
      posX:
        type: integer
      posY:
        type: integer
      roomID:
        type: integer
      row:
        type: string
      type:
        type: string
    type: object
...
  handler.SessionResponseDTO:
    properties:
      id:
        type: integer
      price:
        type: integer
      room_type:
        type: string
      session_type:
        type: string
      start_time:
        type: string
    type: object
  handler.TicketRequest:
    properties:
      seat_id:
        type: integer
      type:
        allOf:
        - $ref: '#/definitions/bookings.TicketType'
        enum:
        - STANDARD
        - HALF
        - FREE
    required:
    - type
    type: object
  handler.TicketResponseDTO:
    properties:
      cinema:
        type: string
      date:
        type: string
      movie_name:
        type: string
      qr_code:
        type: string
      room:
        type: string
      seat:
        type: string
      status:
        type: string
      ticket_id:
        type: string
      user:
        $ref: '#/definitions/handler.UserBookingDTO'
    type: object
...
  handler.UserBookingDTO:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  httputil.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8003
info:
  contact:
    email: suporte@screek.com
    name: Suporte screeK
    url: http://www.screek.com/support
  description: API para a plataforma de cinema screeK.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: screeK API
  version: "1.0"
paths:
...
  /bookings/{id}/sessions:
    get:
      description: Retorna as sessões de um filme agrupadas por cinema
      parameters:
      - description: ID do filme (TMDB)
        in: path
        name: id
        required: true
        type: integer
      - description: Cidade
        in: query
        name: city
        required: true
        type: string
      - description: Data
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.CinemaSessionsResponseDTO'
            type: array
      summary: Consultar sessões de um filme
      tags:
      - Bookings
  /bookings/admin/sessions/{id}/cancel:
    post:
      description: Cancela todos os ingressos de uma sessão em caso de problemas técnicos
        no cinema
      parameters:
      - description: ID da Sessão
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.MessageResponse'
      security:
      - BearerAuth: []
      summary: Cancelar sessão por contingência (Admin)
      tags:
      - Admin
  /bookings/admin/sessions/{id}/tickets:
    get:
      description: Retorna todos os ingressos vendidos para uma sessão específica
      parameters:
      - description: ID da Sessão
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.TicketResponseDTO'
            type: array
      security:
      - BearerAuth: []
      summary: Listar ingressos da sessão (Admin)
      tags:
      - Admin
  /bookings/admin/tickets/{id}/cancel:
    post:
      description: Cancela um ingresso sem restrição de tempo ou dono
      parameters:
      - description: ID do Ticket (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.MessageResponse'
      security:
      - BearerAuth: []
      summary: Cancelar ingresso (Admin)
      tags:
      - Admin
  /bookings/playing:
    get:
      description: Retorna todos os filmes que possuem sessões ativas na cidade e
        data informada
      parameters:
      - description: Cidade do cinema
        in: query
        name: city
        required: true
        type: string
      - description: 'Data (format: YYYY-MM-DD)'
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.MovieResponseDTO'
            type: array
      summary: Listar filmes em cartaz
      tags:
      - Bookings
  /bookings/sessions/{id}/seats:
    get:
      description: |-
        Retorna todos os assentos de uma sessão e seu status de ocupação
        Retorna todos os assentos de uma sessão e seu status de ocupação
      parameters:
      - description: ID da sessão
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.SeatResponseDTO'
            type: array
      summary: Mapa de assentos
      tags:
      - Bookings
      - Bookings
  /bookings/tickets/{id}:
    get:
      description: Retorna informações detalhadas de um ingresso específico
      parameters:
      - description: ID do Ticket (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TicketResponseDTO'
      security:
      - BearerAuth: []
      summary: Detalhe do ingresso
      tags:
      - Bookings
  /bookings/tickets/{id}/cancel:
    post:
      description: Cancela um ingresso e processa estorno se aplicável
      parameters:
      - description: ID do Ticket (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.MessageResponse'
      security:
      - BearerAuth: []
      summary: Cancelar ingresso
      tags:
      - Bookings
  /bookings/tickets/reserve:
    post:
      consumes:
      - application/json
      description: Cria uma reserva temporária de 10 minutos para os assentos selecionados
      parameters:
      - description: Dados da reserva
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ReserveRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.ReserveResponseDTO'
      security:
      - BearerAuth: []
      summary: Reservar assentos
      tags:
      - Bookings
  /bookings/transactions/{id}/pay:
    post:
      consumes:
      - application/json
      description: Processa o pagamento de uma reserva pendente via Stripe
      parameters:
      - description: ID da Transação (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Chave de Idempotência
        in: header
        name: idempotency-key
        required: true
        type: string
      - description: Método de pagamento
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.PayRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.PayResponseDTO'
      security:
      - BearerAuth: []
      summary: Pagar reserva
      tags:
      - Bookings
  /bookings/users/me/tickets:
    get:
      description: Retorna o histórico de ingressos do usuário autenticado
      parameters:
      - description: Filtrar por status (PAID, PENDING, CANCELLED)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.TicketResponseDTO'
            type: array
      security:
      - BearerAuth: []
      summary: Meus ingressos
      tags:
      - Bookings
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"